long to int

+++ fizzbuzz/src/main/java/com/tovk/service/FizzBuzz.java
@@ -2,8 +2,8 @@ package com.tovk.service;
 
 public class FizzBuzz {
   public static String getFizzBuzz(long number) {
-    long divBy3 = number % 3;
-    long divBy5 = number % 5;
+    int divBy3 = (int) (number % 3);
+    int divBy5 = (int) (number % 5);
     if (divBy3 != 0 && divBy5 != 0) {
       return String.valueOf(number);
     } else {
